% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{assayExists}
\alias{assayExists}
\title{check if the given assay already exists within the object}
\usage{
assayExists(fds, assayName)
}
\description{
check if the given assay already exists within the object
}
