% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{assayNames,FraserDataSet-method}
\alias{assayNames,FraserDataSet-method}
\title{Returns the assayNames of FRASER}
\usage{
\S4method{assayNames}{FraserDataSet}(x)
}
\arguments{
\item{x}{FraserDataSet}
}
\value{
Character vector
}
\description{
Returns the assayNames of FRASER
}
