% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{bamFile,FraseRDataSet-method}
\alias{bamFile,FraseRDataSet-method}
\alias{bamFile<-,FraseRDataSet-method}
\title{Get/Set the bamFile}
\usage{
\S4method{bamFile}{FraseRDataSet}(object)

\S4method{bamFile}{FraseRDataSet}(object) <- value
}
\arguments{
\item{object}{A FraseRDataSet object.}
}
\value{
A \code{vector} with the bamFile for each sample
}
\description{
Get/Set the bamFile
}
\examples{
settings <- createTestFraseRSettings()
bamFile(settings)
bamFile(settings) <- file.path("bamfiles", samples(settings), "rna-seq.bam")
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
