% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRSettings-class.R
\docType{methods}
\name{bamFiles,FraseRSettings-method}
\alias{bamFiles,FraseRSettings-method}
\alias{bamFiles<-,FraseRSettings-method}
\title{Get/Set the bamFiles from the FraseRSettings object}
\usage{
\S4method{bamFiles}{FraseRSettings}(object)

\S4method{bamFiles}{FraseRSettings}(object) <- value
}
\arguments{
\item{object}{A FraseRSettings object.}
}
\value{
A \code{vector} with the bamFiles for each sample
}
\description{
Get/Set the bamFiles from the FraseRSettings object
}
\examples{
settings <- createTestFraseRSettings()
bamFiles(settings)
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
