% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistributions.R
\name{breakTies}
\alias{breakTies}
\title{breaks ties in a qq plot to get a better distributed p-value plot}
\usage{
breakTies(x, logBase = 10, decreasing = TRUE)
}
\description{
breaks ties in a qq plot to get a better distributed p-value plot
}
