% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePSIValue.R
\name{calculateDeltaPsiValue}
\alias{calculateDeltaPsiValue}
\title{calculates the delta psi value and stores it as an assay}
\usage{
calculateDeltaPsiValue(fds, psiType)
}
\description{
calculates the delta psi value and stores it as an assay
}
