% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStats.R
\name{calculatePValues}
\alias{calculatePValues}
\title{calculates the P-Value for the given FraseR dataset object
The P-Value calculation is based on the given method in the
FraseRSettings object}
\usage{
calculatePValues(dataset, internBPPARAM = SerialParam())
}
\description{
calculates the P-Value for the given FraseR dataset object
The P-Value calculation is based on the given method in the
FraseRSettings object
}
\examples{
  fds <- countRNAData(createTestFraseRSettings())
  fds <- calculatePValues(fds)
}
