% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePSIValue.R
\name{calculateSitePSIValue}
\alias{calculateSitePSIValue}
\title{This function calculates the site PSI values for each splice site
based on the FraseRDataSet object}
\usage{
calculateSitePSIValue(dataset)
}
\description{
This function calculates the site PSI values for each splice site
based on the FraseRDataSet object
}

