% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStats.R
\name{calculateZScores}
\alias{calculateZScores}
\title{Calculate the zscore for each psi value.}
\usage{
calculateZScores(dataset)
}
\description{
Calculate the zscore for each psi value.
}
\examples{
  fds <- counRNAData(createTestFraseRSettings())
  fds <- calculatePSIValues(fds)
  fds <- calculateZScores(fds)
}
