% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{checkFraseRDataSet}
\alias{checkFraseRDataSet}
\alias{checkCountData}
\title{Input check functions}
\usage{
checkFraseRDataSet(fds)

checkCountData(fds, stop = TRUE)
}
\description{
Checks all user input and returns corresponding messages
checkFraseRDataSet
}
