% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{checkSeqLevelStyle}
\alias{checkSeqLevelStyle}
\title{Map between individual seq level style and dataset common one
for counting and aggregating the reads}
\usage{
checkSeqLevelStyle(gr, fds, sampleID, sampleSpecific = FALSE)
}
\description{
Map between individual seq level style and dataset common one
for counting and aggregating the reads
}
