% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{cleanCache}
\alias{cleanCache}
\title{clear the files in the cache to start fresh}
\usage{
cleanCache(fds, all = FALSE, cache = TRUE, assays = FALSE,
  results = FALSE)
}
\description{
clear the files in the cache to start fresh
}
\examples{
    fds <- createTestFraseRSettings()
    cleanCache(fds)

}
