% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{condition}
\alias{condition}
\alias{condition<-}
\alias{condition,FraseRDataSet-method}
\alias{condition<-,FraseRDataSet-method}
\title{Get the condition for each sample}
\usage{
condition(object)

condition(object) <- value

\S4method{condition}{FraseRDataSet}(object)

\S4method{condition}{FraseRDataSet}(object) <- value
}
\arguments{
\item{object}{A FraseRDataSet object.}

\item{value}{A vector with the conditions that should be assigned.}
}
\value{
A \code{vector} with the condition per sample
    If the condition column is not set sequence of numbers is returned.
}
\description{
Get the condition for each sample
}
\examples{
fds <- createTestFraseRSettings()
condition(fds)
condition(fds) <- 1:dim(fds)[2]
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
