% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_script.R
\name{count_rna}
\alias{count_rna}
\title{Counting the split reads and the non spliced reads
in the given bamfile list. Can be computed in parallel}
\usage{
count_rna(bamfiles, BPPARAM)
}
\arguments{
\item{bamfiles}{a list of bamfiles}

\item{BPPARAM}{the BiocParallel param object to run jobs in parallel}
}
\value{
SummarizedExperiment count data
}
\description{
Counting the split reads and the non spliced reads
in the given bamfile list. Can be computed in parallel
}
\examples{
  count_rna("./data/my.bam")

}

