% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countRNAseqData.R
\name{countRNAData}
\alias{countRNAData}
\title{Count RNA-seq data}
\usage{
countRNAData(settings, internBPPARAM = SerialParam(),
  globalJunctionMap = NULL)
}
\arguments{
\item{settings}{A FraseRSetting object with all the information
how and what to count}

\item{internBPPARAM}{A BiocParallel param object to configure the
parallel backend of the internal loop}

\item{globalJunctionMap}{A object or file containing a map of
all junction of interest across all samples}
}
\value{
FraseRDataSet
}
\description{
This method extracts and counts the split reads and
            non spliced reads from a RNA bam file
}
\details{
TODO
}
\examples{
  counRNAData(createTestFraseRSettings())
}
