% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{counts<-,FraseRDataSet,ANY-method}
\alias{counts<-,FraseRDataSet,ANY-method}
\title{setter for count data}
\usage{
\S4method{counts}{FraseRDataSet,ANY}(object, type = NULL,
  side = c("ofInterest", "otherSide")) <- value
}
\description{
setter for count data
}
