% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{createTestFraseRSettings}
\alias{createTestFraseRSettings}
\title{Create a test case dataset (sample information only)
to be used in the vignette and to explore the
functionallity of the FraseR package.}
\usage{
createTestFraseRSettings()
}
\value{
a FraseRSettings object which contains a test case
}
\description{
Create a test case dataset (sample information only)
to be used in the vignette and to explore the
functionallity of the FraseR package.
}
\examples{
fds <- createTestFraseRSettings()
fds

}
