% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNSetterFuns.R
\name{currentType}
\alias{currentType}
\alias{currentType<-}
\title{Set/get psi type}
\usage{
currentType(fds)

currentType(fds) <- value
}
\arguments{
\item{fds}{FraseRDataSet}

\item{value}{If given, is sets the psi type. It is required to be a scalar 
character.}
}
\value{
character() (getter) or FraseRDataSet(setter)
}
\description{
Set and returns the psi type that is to be used within several methods in 
the FraseR package.
}
\examples{
  # get data
  fds <- makeSimulatedFraserDataSet(m=50, j=1000)

  # set
  currentType(fds) <- "psi5"

  # get
  currentType(fds)

}
