% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{samples}
\alias{samples}
\alias{samples<-}
\alias{condition}
\alias{condition<-}
\alias{bamFile}
\alias{bamFile<-}
\alias{name}
\alias{name<-}
\alias{strandSpecific}
\alias{strandSpecific<-}
\alias{workingDir}
\alias{workingDir<-}
\alias{scanBamParam}
\alias{scanBamParam<-}
\alias{nonSplicedReads}
\alias{nonSplicedReads<-}
\alias{samples,FraseRDataSet-method}
\alias{samples<-,FraseRDataSet-method}
\alias{condition,FraseRDataSet-method}
\alias{condition<-,FraseRDataSet-method}
\alias{bamFile,FraseRDataSet-method}
\alias{bamFile<-,FraseRDataSet-method}
\alias{name,FraseRDataSet-method}
\alias{name<-,FraseRDataSet-method}
\alias{workingDir,FraseRDataSet-method}
\alias{workingDir<-,FraseRDataSet-method}
\alias{strandSpecific,FraseRDataSet-method}
\alias{strandSpecific<-,FraseRDataSet-method}
\alias{scanBamParam,FraseRDataSet-method}
\alias{scanBamParam<-,FraseRDataSet-method}
\alias{nonSplicedReads,FraseRDataSet-method}
\alias{nonSplicedReads<-,FraseRDataSet-method}
\alias{FraseR.mcols.get}
\alias{FraseR.rowRanges.get}
\alias{mapSeqlevels}
\title{Getter/Setter methods for the FraseRDataSet}
\usage{
samples(object)

samples(object) <- value

condition(object)

condition(object) <- value

bamFile(object)

bamFile(object) <- value

name(object)

name(object) <- value

strandSpecific(object)

strandSpecific(object) <- value

workingDir(object)

workingDir(object) <- value

scanBamParam(object)

scanBamParam(object) <- value

nonSplicedReads(object)

nonSplicedReads(object) <- value

\S4method{samples}{FraseRDataSet}(object)

\S4method{samples}{FraseRDataSet}(object) <- value

\S4method{condition}{FraseRDataSet}(object)

\S4method{condition}{FraseRDataSet}(object) <- value

\S4method{bamFile}{FraseRDataSet}(object)

\S4method{bamFile}{FraseRDataSet}(object) <- value

\S4method{name}{FraseRDataSet}(object)

\S4method{name}{FraseRDataSet}(object) <- value

\S4method{workingDir}{FraseRDataSet}(object)

\S4method{workingDir}{FraseRDataSet}(object) <- value

\S4method{strandSpecific}{FraseRDataSet}(object)

\S4method{strandSpecific}{FraseRDataSet}(object) <- value

\S4method{scanBamParam}{FraseRDataSet}(object)

\S4method{scanBamParam}{FraseRDataSet}(object) <- value

\S4method{nonSplicedReads}{FraseRDataSet}(object)

\S4method{nonSplicedReads}{FraseRDataSet}(object) <- value

FraseR.mcols.get(x, type = NULL, ...)

FraseR.rowRanges.get(x, type = NULL, ...)

mapSeqlevels(fds, style = "UCSC", ...)
}
\arguments{
\item{object}{A FraseRDataSet object.}

\item{value}{The new value that should replace the current one.}

\item{x}{A FraseRDataSet object.}

\item{type}{The psi type (psi3, psi5 or psiSite)}

\item{...}{Further parameters. For mapSeqLevels: further parameters 
passed to GenomeInfoDb::mapSeqlevels().}

\item{fds}{FraseRDataSet}

\item{style}{The style of the chromosome names.}
}
\value{
Getter method return the respective current value.
}
\description{
The following methods are getter and setter methods to extract or set
certain values of a FraseRDataSet object.
}
\details{
\code{samples} sets or gets the sample IDs; \code{condition} ;
\code{}
\code{nonSplicedReads} return a RangedSummarizedExperiment object 
containing the counts for the non spliced reads overlapping splice 
sites in the fds.
\code{}
}
\examples{
fds <- createTestFraseRDataSet()
samples(fds)
samples(fds) <- 1:dim(fds)[2]
condition(fds)
condition(fds) <- 1:dim(fds)[2]
bamFile(fds) # file.paths or objects of class BamFile
bamFile(fds) <- file.path("bamfiles", samples(fds), "rna-seq.bam")
name(fds)
name(fds) <- "My Analysis"
workingDir(fds)
workingDir(fds) <- tempdir()
strandSpecific(fds)
strandSpecific(fds) <- TRUE
strandSpecific(fds) <- "reverse"
strandSpecific(fds)
scanBamParam(fds)
scanBamParam(fds) <- ScanBamParam(mapqFilter=30)
nonSplicedReads(fds)
rowRanges(fds)
rowRanges(fds, type="psiSite")
mcols(fds, type="psi5")
mcols(fds, type="psiSite")
seqlevels(fds)
seqlevels(mapSeqlevels(fds, style="UCSC"))
seqlevels(mapSeqlevels(fds, style="Ensembl"))
seqlevels(mapSeqlevels(fds, style="dbSNP"))

}
\author{
Christian Mertes \email{mertes@in.tum.de}

Christian Mertes \email{mertes@in.tum.de}

Christian Mertes \email{mertes@in.tum.de}
}
