% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{samples}
\alias{samples}
\alias{samples<-}
\alias{condition}
\alias{condition<-}
\alias{bamFile}
\alias{bamFile<-}
\alias{name}
\alias{name<-}
\alias{strandSpecific}
\alias{strandSpecific<-}
\alias{workingDir}
\alias{workingDir<-}
\alias{scanBamParam}
\alias{scanBamParam<-}
\alias{nonSplicedReads}
\alias{nonSplicedReads<-}
\title{Getters/Setters for FraseRDataSet values}
\usage{
samples(object)

samples(object) <- value

condition(object)

condition(object) <- value

bamFile(object)

bamFile(object) <- value

name(object)

name(object) <- value

strandSpecific(object)

strandSpecific(object) <- value

workingDir(object)

workingDir(object) <- value

scanBamParam(object)

scanBamParam(object) <- value

nonSplicedReads(object)

nonSplicedReads(object) <- value
}
\arguments{
\item{object}{FraseRDataSet}

\item{value}{The new value.}
}
\value{
The value assigned to the FraseRDataSet object.
}
\description{
Getters/Setters for FraseRDataSet values
}
