% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNSetterFuns.R
\name{featureExclusionMask}
\alias{featureExclusionMask}
\alias{featureExclusionMask,}
\alias{`featureExclusionMask<-`}
\alias{featureExclusionMask<-}
\title{Feature exclusion}
\usage{
featureExclusionMask(fds)

featureExclusionMask(fds) <- value
}
\arguments{
\item{value}{A logical vector of the length of the features. If \code{TRUE},
the corresponding feature will be excluded from the encoding dimension
fit.}

\item{ods}{An OutriderDataSet object}
}
\value{
The exclusion vector
}
\description{
To remove certain junctions from being used in the train step of the encoding dimension
we can set the \code{featureExclusion} vector to \code{FALSE}. This can be helpfull if
we have local linkage between features which we do not want to model by the autoencoder.
}
\examples{
ods <- makeExampleFraseRDataSet()
featureExclusionMask(fds) <- sample(c(FALSE, TRUE), nrow(fds), replace=TRUE)

featureExclusionMask(fds)

}
