% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterExpression.R
\name{filterExpression}
\alias{filterExpression}
\title{Filtering FraseRDataSets}
\usage{
filterExpression(fds, minExpressionInOneSample = 5, quantile = 0.7,
  quantileMinExpression = 1, filter = FALSE)
}
\description{
Filtering FraseRDataSets
}
