% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationOfRanges.R
\name{findAnnotatedJunction}
\alias{findAnnotatedJunction}
\title{Find annotated junctions}
\usage{
findAnnotatedJunction(fds, annotation, annotateNames = TRUE,
  seqLevelStyle = seqlevelsStyle(fds)[1],
  stranded = strandSpecific(fds), ...)
}
\value{
FraseRDataSet
}
\description{
Annotate the object with a given annotation.
}
