% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_encoding_dimensions.R
\name{fit_autoenc}
\alias{fit_autoenc}
\title{### MODEL}
\usage{
fit_autoenc(fds, type = currentType(fds), q_guess = round(ncol(fds)/4),
  correction, noiseRatio = 0.5, BPPARAM = bpparam(), iterations = 3,
  nrDecoderBatches = 1, verbose = FALSE)
}
\description{
### MODEL
}
