% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_autoencoder.r
\name{fit_keras_bb_dea}
\alias{fit_keras_bb_dea}
\title{Keras denoising autoencoder with betabinomial loss}
\usage{
fit_keras_bb_dea(fds, q, type, noiseAlpha, rhoRange, BPPARAM, lr = 1e-04,
  epochs = 50, patience = 5, reUseWeights = TRUE, iterations = 10)
}
\description{
Install environment: (bash commands)
}
\details{
envName=omicsOUTRIDER
  module load i12g/anaconda/3-5.0.1
  conda create -n $envName python=3.6
  /opt/modules/i12g/anaconda/3-5.0.1/envs/$envName/bin/pip install -U keras tensorflow jupyterlab

To use it run before training:

library(keras)
  use_python("/opt/modules/i12g/anaconda/3-5.0.1/envs/omicsOUTRIDER/bin/python", required=TRUE)
}
