% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{FraseR.mcols.get}
\alias{FraseR.mcols.get}
\title{getter and setter for mcols}
\usage{
FraseR.mcols.get(x, type = NULL, ...)
}
\value{
mcols
}
\description{
getter and setter for mcols
}
\examples{
  fds <- makeExampleFraseRDataSet()
  mcols(fds, type="psi5")
}
