% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseR-package.R, R/Fraser-pipeline.R
\name{FraseR}
\alias{FraseR}
\alias{FraseR}
\title{FraseR: A package providing a workflow to detect aberrant splicing events
    in RNA sequencing data in a rare disease cohort}
\usage{
FraseR(settings, internBPPARAM = SerialParam())
}
\arguments{
\item{settings}{A FraseRSetting object with all the information
how and what to count}

\item{internBPPARAM}{A BiocParallel param object to configure the
parallel backend of the internal loop for counting}
}
\value{
FraseRDataSet
}
\description{
FraseR: A package providing a workflow to detect aberrant splicing events
    in RNA sequencing data in a rare disease cohort

This function performs a default analysis of RNA-seq data
}
\examples{
  fds <- FraseR(createTestFraseRSettings())
  plotSampleResults(fds)
}
