% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{FraseR.results}
\alias{FraseR.results}
\title{obtain the results for the given analysis pipeline}
\usage{
FraseR.results(x, sampleIDs = samples(x), fdrCut = 0.05, zscoreCut = 2,
  psiType = c("psi3", "psi5", "psiSite"), redo = FALSE)
}
\description{
obtain the results for the given analysis pipeline
}
