% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{FraseR.rowRanges.get}
\alias{FraseR.rowRanges.get}
\title{Getter and setter for rowRanges}
\usage{
FraseR.rowRanges.get(x, type = NULL, ...)
}
\arguments{
\item{x}{FraseRDataSet}

\item{type}{The psi type.}

\item{...}{Parameters passed on to rowRanges()}
}
\value{
GRanges object

FraseRDataSet
}
\description{
Getter and setter for rowRanges
}
\examples{
  fds <- createTestFraseRDataSet()
  
  rowRanges(fds)
  rowRanges(fds, type="psiSite")

}
