% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRDataSet-class.R
\docType{class}
\name{FraseRDataSet-class}
\alias{FraseRDataSet-class}
\title{FraseRDataSet}
\description{
This class is designed to store the whole FraseR data set
needed for an analysis of a disease cohort
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}

