% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRDataSet-class.R
\name{FraseRDataSet}
\alias{FraseRDataSet}
\title{The constructor function for FraseRSettings}
\usage{
FraseRDataSet(...)
}
\arguments{
\item{...}{Any parameters corresponding to the slots and their possible
values. See \linkS4class{FraseRDataSet}}
}
\value{
A FraseRDataSet object.
}
\description{
The constructor function for FraseRSettings
}
\examples{
    fraser <- FraseRDataSet()
    fraser <- countRNAData(createTestFraseRSettings())
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}

