% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRSettings-class.R
\docType{class}
\name{FraseRSettings-class}
\alias{FraseRSettings-class}
\title{FraseRSettings}
\description{
This class is designed to store the sample annotation 
and settings for the computation of the FraseR package
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
