% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRSettings-class.R
\name{FraseRSettings}
\alias{FraseRSettings}
\title{The constructor function for FraseRSettings}
\usage{
FraseRSettings(...)
}
\arguments{
\item{...}{Any parameters corresponding to the slots and their possible
values. See \linkS4class{FraseRSettings}}
}
\value{
A FraseRSettings object.
}
\description{
The constructor function for FraseRSettings
}
\examples{
    settings <- FraseRSettings()
    settings <- FraseRSettings(fread(system.file(
            "extdata", "sampleTable.tsv", package="FraseR", mustWork=TRUE
    )))
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
