% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{FraseRShiny}
\alias{FraseRShiny}
\title{Present the FraseR results in a shiny app}
\usage{
FraseRShiny(fds, fdsres = NULL, options = list(), server = !interactive())
}
\description{
Present the FraseR results in a shiny app
}
\examples{
    fds <- createTestFraseRDataSet()
    # for interactive sessions
    FraseRShiny(fds)

    # for running a shiny application as server
    FraseRShiny(fds, server=TRUE)

}
