% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{FraseRShinyApp}
\alias{FraseRShinyApp}
\title{Present the FraseR results as shiny app}
\usage{
FraseRShinyApp(fds, fdsres = NULL, server = !interactive(), ...)
}
\value{
shiny application
}
\description{
Present the FraseR results as shiny app
}
\examples{
    fds <- createTestFraseRDataSet()

    # for interactive sessions
    myShinyApp <- FraseRShinyApp(fds)

    # for running a shiny application as server
    myShinyAppObj <- FraseRShinyApp(fds, server=TRUE)

}
