% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSimulatedDataset.R
\name{makeSimulatedFraserDataSet_BetaBinomial}
\alias{makeSimulatedFraserDataSet_BetaBinomial}
\title{Create an simulated example data set for FraseR}
\usage{
makeSimulatedFraserDataSet_BetaBinomial(m = 200, j = 10000, q = 10)
}
\arguments{
\item{m}{Number of simulated samples}

\item{j}{Number of simulated junctions}

\item{q}{number of simulated latent variables.}
}
\value{
An FraserDataSet containing an example dataset based on 
        simulated data
}
\description{
Simulates a data set based on random counts following a 
negative binomial distribution with injected outliers with 
a fixed z score away from the mean of the gene.
}
\examples{
# A generic dataset 
fds1 <- makeSimulatedFraserDataSet()
fds1
fds1 <- injectOutliers(fds1, type="psi3", deltaPSI=0.4, freq=1E-3, inj="both")

# A generic dataset with specificed sample size and injection method
fds2 <- makeSimulatedFraserDataSet(m=100, j=500, inj='low')
fds2

}
