% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{mcols,FraseRDataSet-method}
\alias{mcols,FraseRDataSet-method}
\title{getter and setter for mcols}
\usage{
\S4method{mcols}{FraseRDataSet}(x, type = NULL, ...)
}
\description{
getter and setter for mcols
}
