% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFds.R
\name{mergeFDS}
\alias{mergeFDS}
\title{Merge two FraserDataSet objects with different splice sites and junctions}
\usage{
mergeFDS(fds1, fds2, join_type = c("outer", "inner"), fds_name = "merged_fds")
}
\arguments{
\item{fds1}{First FraserDataSet object.}

\item{fds2}{Second FraserDataSet object.}

\item{join_type}{Type of join for rowRanges: "outer" (default) or "inner".}

\item{fds_name}{Name for the merged FraserDataSet.}
}
\value{
A merged FraserDataSet object containing combined samples and assays.
}
\description{
This function merges two FraserDataSet objects that have different
splice sites and junctions but no overlapping samples. It combines
only the rawCountsJ and rawCountsSS assays and recalculates startID
and endID for the merged rowRanges.
}
\examples{
# merged_fds <- mergeFDS(fds1, fds2, join_type="outer")
}
