% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRSettings-class.R
\docType{methods}
\name{method,FraseRSettings-method}
\alias{method,FraseRSettings-method}
\alias{method<-,FraseRSettings-method}
\title{Get/Set the statistical method to use for P-value calculation
 from the FraseRSettings object}
\usage{
\S4method{method}{FraseRSettings}(object)

\S4method{method}{FraseRSettings}(object) <- value
}
\arguments{
\item{object}{A FraseRSettings object.}
}
\value{
A character string representing the statistical method
}
\description{
Get/Set the statistical method to use for P-value calculation
 from the FraseRSettings object
}
\examples{
settings <- createTestFraseRSettings()
method(settings)
method(settings) <- "betaBinomial"
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
