% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_script.R
\name{my_first_function}
\alias{my_first_function}
\title{This is my first function and some doc}
\usage{
my_first_function(arg = TRUE)
}
\arguments{
\item{arg}{this is a dummy parameter}
}
\value{
NULL nothing
}
\description{
This is my first function and some doc
}
\examples{
  my_first_function()

}

