% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{na2default}
\alias{na2default}
\alias{na2false}
\alias{na2zero}
\alias{null2na}
\alias{null2default}
\title{Convert to default values}
\usage{
na2default(x, default = FALSE)

null2default(x, default = NA)

na2false(x)

na2zero(x)

null2na(x)
}
\value{
vector
}
\description{
Convert all NA's of a input vector or of a
single dimension matrix/data.table to FALSE
}
\details{
Convert NULL to NA or to another default value
}
\examples{
  a <- c(TRUE, FALSE, NA, TRUE, NA)
  na2false(a)

  dt <- data.table(a)
  na2false(dt)

  null2na(NULL)
  null2na(1:10)

}
