% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{na2false}
\alias{na2false}
\title{convert all NA's of a input vector or of a
single dimension matrix/data.table to FALSE}
\usage{
na2false(x)
}
\description{
convert all NA's of a input vector or of a
single dimension matrix/data.table to FALSE
}
\examples{
  a <- c(TRUE, FALSE, NA, TRUE, NA)
  na2false(a)

  dt <- data.table(a)
  na2false(dt)

}
