% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{name}
\alias{name}
\alias{name<-}
\alias{name,FraseRDataSet-method}
\alias{name<-,FraseRDataSet-method}
\title{Get/Set the name of the analysis}
\usage{
name(object)

name(object) <- value

\S4method{name}{FraseRDataSet}(object)

\S4method{name}{FraseRDataSet}(object) <- value
}
\arguments{
\item{object}{A FraseRDataSet object.}

\item{value}{The new name of the analysis.}
}
\value{
A character string representing the name of the analysis
}
\description{
Get/Set the name of the analysis
}
\examples{
settings <- createTestFraseRSettings()
name(settings)
name(settings) <- "My Analysis"
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
