% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{nonSplicedReads,FraseRDataSet-method}
\alias{nonSplicedReads,FraseRDataSet-method}
\alias{nonSplicedReads<-,FraseRDataSet-method}
\title{accessor for the non spliced reads object within the FraseRDataSet object}
\usage{
\S4method{nonSplicedReads}{FraseRDataSet}(object)

\S4method{nonSplicedReads}{FraseRDataSet}(object) <- value
}
\description{
accessor for the non spliced reads object within the FraseRDataSet object

setter for the non spliced reads object within the FraseRDataSet object
}
