% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_encoding_dimensions.R
\name{optimHyperParams}
\alias{optimHyperParams}
\title{Find optimal encoding dimension}
\usage{
optimHyperParams(fds, type, correction, q_param = seq(2, min(40,
  ncol(fds)), by = 3), noise_param = c(0, 0.5, 1, 2, 5),
  minDeltaPsi = 0.1, iterations = 5, setSubset = 15000,
  injectFreq = 0.01, nrDecoderBatches = 1, BPPARAM = bpparam(),
  internalThreads = 3)
}
\arguments{
\item{fds}{A FraseRDataSet object}

\item{type}{The type of PSI (psi5, psi3 or psiSite for theta/splicing 
efficiency)}

\item{correction}{The method that should be used to correct for confounders.}

\item{minDeltaPsi}{}

\item{iterations}{The maximal number of iterations. When the autoencoder has 
not yet converged after these number of iterations, the fit stops anyway.}

\item{BPPARAM}{the BiocParallel parameters for the parallelization}
}
\value{
FraseRDataSet
}
\description{
Finds the optimal encoding dimension by injecting artificial splicing outlier
ratios while maximizing the precision-recall curve.
}
\examples{
  # generate data
  fds <- makeSimulatedFraserDataSet(j=500, m=20, q=4)
  
  # run hyperparameter optimization
  fds <- optimHyperParams(fds, type="psi5", correction="PCA")
  
  # get estimated optimal dimension of the latent space
  bestQ(fds, type="psi5")
  
  # plot the AUC for the different encoding dimensions tested
  plotEncDimSearch(fds, type="psi5")

}
