% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRSettings-class.R
\docType{methods}
\name{outputFolder,FraseRSettings-method}
\alias{outputFolder,FraseRSettings-method}
\alias{outputFolder<-,FraseRSettings-method}
\title{Get/Set the output folder from the FraseRSettings object}
\usage{
\S4method{outputFolder}{FraseRSettings}(object)

\S4method{outputFolder}{FraseRSettings}(object) <- value
}
\arguments{
\item{object}{A FraseRSettings object.}
}
\value{
A path
}
\description{
Get/Set the output folder from the FraseRSettings object
}
\examples{
settings <- createTestFraseRSettings()
outputFolder(settings)
outputFolder(settings) <- tempdir()

}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
