% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRSettings-class.R
\docType{methods}
\name{parallel,FraseRSettings-method}
\alias{parallel,FraseRSettings-method}
\alias{parallel<-,FraseRSettings-method}
\title{Get/Set the parallel param object from the FraseRSettings object}
\usage{
\S4method{parallel}{FraseRSettings}(object)

\S4method{parallel}{FraseRSettings}(object) <- value
}
\arguments{
\item{object}{A FraseRSettings object.}
}
\value{
A parallel param object
}
\description{
Get/Set the parallel param object from the FraseRSettings object
}
\examples{
settings <- createTestFraseRSettings()
parallel(settings)
parallel(settings) <- SerialParam()
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
