% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistributions.R
\name{plotCountCorHeatmap}
\alias{plotCountCorHeatmap}
\title{Plot count correlation}
\usage{
plotCountCorHeatmap(fds, type = c("psi5", "psi3", "psiSite"),
  logit = FALSE, topN = 50000, minMedian = 1, main = NULL,
  show_rownames = FALSE, show_colnames = FALSE, annotation_col = NA,
  annotation_row = NA, ...)
}
\description{
Plot count correlation
}
