% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotCountsAtSite}
\alias{plotCountsAtSite}
\title{plot count distribution}
\usage{
plotCountsAtSite(gr, fds, type, sampleIDs = NULL, plotLegend = TRUE,
  data = NULL, zeroVal = 0.64)
}
\description{
plot count distribution
}
