% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{plotExpectedVsObservedPsi}
\alias{plotExpectedVsObservedPsi}
\title{Expected over Overserved plot}
\usage{
plotExpectedVsObservedPsi(fds, type = c("psi5", "psi3", "psiSite"),
  idx = NULL, result = NULL, colGroup = NULL, main = NULL,
  basePlot = TRUE, padjCutoff = 0.05)
}
\description{
Plots the expected psi value over the observed psi value of the given
junction.
}
