% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{plotExpression}
\alias{plotExpression}
\title{Junction expression plot}
\usage{
plotExpression(fds, type = c("psi5", "psi3", "psiSite"), site = NULL,
  result = NULL, colGroup = NULL, basePlot = TRUE,
  title = paste0("Expression plot: ", site), ...)
}
\description{
Plots the observed split reads of the junction of interest over all reads
coming from the given donor/acceptor.
}
