% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotSampleResults}
\alias{plotSampleResults}
\title{Plot the results of the FraseR analysis pipeline
All three types are plotted into one HTML file
based on plotly.}
\usage{
plotSampleResults(fds, sampleID = NULL, file = NULL, dir = NULL,
  browseIt = FALSE, BPPARAM = bpparam())
}
\value{
The html file with the plots.
}
\description{
Plot the results of the FraseR analysis pipeline
All three types are plotted into one HTML file
based on plotly.
}
\examples{
# TODO
TODO <- 1
#     fds <- FraseR()
#     plotSampleResults(fds, "sample1")
#     plotSampleResults(fds, "sample1", "result.html")
}
