% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistributions.R
\name{plotValueVsCounts}
\alias{plotValueVsCounts}
\title{plot a given value against the counts}
\usage{
plotValueVsCounts(gr, fds, type, sample = NULL, delta = FALSE,
  plotLog = FALSE, rmZeroCts = FALSE, data = NULL, ...)
}
\description{
plot a given value against the counts
}
