% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot}
\usage{
plotVolcano(fds, sampleID, type, minDeltaPsi = 0.3, padjCutoff = 0.05,
  basePlot = TRUE, aggregate = FALSE, main = paste0("Volcano plot: ",
  sampleID))
}
\description{
Plots the p values over the delta psi values, known as volcano plot.
Visualizes per sample the outliers. By type and aggregate by
gene if requested.
}
