% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{psiTypes_avail}
\alias{psiTypes_avail}
\alias{psiTypes}
\title{Available splice metrics}
\format{
An object of class \code{character} of length 4.

An object of class \code{character} of length 1.
}
\usage{
psiTypes_avail

psiTypes
}
\description{
Available splice metrics
}
\section{Functions}{
\itemize{
\item \code{psiTypes}: Splice metrics that are run by default
}}

\examples{
  # to show available splice metrics:
  psiTypes_avail

  # to show splice metrics selected to be fitted:
  psiTypes

}
\keyword{datasets}
