% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{psiTypes_avail}
\alias{psiTypes_avail}
\title{Available splice metrics}
\format{
An object of class \code{character} of length 4.
}
\usage{
psiTypes_avail
}
\description{
Available splice metrics
}
\examples{
  # to show available splice metrics:
  psiTypes_avail

}
\keyword{datasets}
