% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{results}
\alias{results}
\alias{results,FraseRDataSet-method}
\title{Extracting results}
\usage{
results(x, ...)

\S4method{results}{FraseRDataSet}(x, sampleIDs = samples(x),
  fdrCut = 0.1, zscoreCut = 2, dPsiCut = 0.03, psiType = c("psi3",
  "psi5", "psiSite"), BPPARAM = bpparam())
}
\description{
The result function extracts the results from the given analysis object
based on the given options and cutoffs.
}
