% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRSettings-class.R
\docType{methods}
\name{sampleData,FraseRSettings-method}
\alias{sampleData,FraseRSettings-method}
\alias{sampleData<-,FraseRSettings-method}
\title{Get/Set the sampleData table the FraseRSettings object}
\usage{
\S4method{sampleData}{FraseRSettings}(object)

\S4method{sampleData}{FraseRSettings}(object) <- value
}
\arguments{
\item{object}{A FraseRSettings object.}
}
\value{
A \code{data.table} with all sample infromation (id, bamfile, group)
}
\description{
Get/Set the sampleData table the FraseRSettings object
}
\examples{
settings <- createTestFraseRSettings()
sampleData(settings)
sampleData(settings) <- 1:length(settings)
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
