% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRSettings-class.R
\docType{methods}
\name{sampleGroup,FraseRSettings-method}
\alias{sampleGroup,FraseRSettings-method}
\alias{sampleGroup<-,FraseRSettings-method}
\title{Get the group definition per sample based on the
\code{sampleData} table slot}
\usage{
\S4method{sampleGroup}{FraseRSettings}(object)

\S4method{sampleGroup}{FraseRSettings}(object) <- value
}
\arguments{
\item{object}{A FraseRSettings object.}
}
\value{
A \code{vector} with the group identifiers per sample object
}
\description{
Get the group definition per sample based on the
\code{sampleData} table slot
}
\examples{
settings <- createTestFraseRSettings()
sampleGroup(settings)
sampleGroup(settings) <- 1:length(settings)
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
