% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRSettings-class.R
\docType{methods}
\name{samples,FraseRSettings-method}
\alias{samples,FraseRSettings-method}
\alias{samples<-,FraseRSettings-method}
\title{Get/Set the sampleIDs from the FraseRSettings object}
\usage{
\S4method{samples}{FraseRSettings}(object)

\S4method{samples}{FraseRSettings}(object) <- value
}
\arguments{
\item{object}{A FraseRSettings object.}
}
\value{
A \code{vector} with all sample IDs
}
\description{
Get/Set the sampleIDs from the FraseRSettings object
}
\examples{
settings <- createTestFraseRSettings()
samples(settings)
samples(settings) <- 1:length(settings)
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
