% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveHDF5Objects.R
\name{saveFraseRDataSet}
\alias{saveFraseRDataSet}
\title{Saves the FraseRDataSet object on disk under the given working dir.
It furthermore uses HDF5 to save all internal assays}
\usage{
saveFraseRDataSet(fds, dir = NULL)
}
\arguments{
\item{fds}{A FraseRDataSet object ot be saved}

\item{dir}{a directory name where to save the objects
(replaces the working directory)}
}
\description{
After saving this can be loaded again with the corresnonding load function
}
\examples{
    fds <- countRNAData(createTestFraseRSettings())
    saveFraseRDataSet(fds)

}
