% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRSettings-class.R
\docType{methods}
\name{scanBamParam,FraseRSettings-method}
\alias{scanBamParam,FraseRSettings-method}
\alias{scanBamParam<-,FraseRSettings-method}
\title{Get/Set the ScanBamParam object from the FraseRSettings object}
\usage{
\S4method{scanBamParam}{FraseRSettings}(object)

\S4method{scanBamParam}{FraseRSettings}(object) <- value
}
\arguments{
\item{object}{A FraseRSettings object.}
}
\value{
A ScanBamParam object
}
\description{
Get/Set the ScanBamParam object from the FraseRSettings object
}
\examples{
settings <- createTestFraseRSettings()
scanBamParam(settings)
scanBamParam(settings) <- ScanBamParam()

}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
