% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultAnnotations.R
\name{spliceTypeAnnotations}
\alias{spliceTypeAnnotations}
\alias{annotateIntronReferenceOverlap}
\alias{annotateSpliceEventType}
\alias{flagBlacklistRegions}
\title{Additional result annotations}
\usage{
annotateIntronReferenceOverlap(fds, txdb, BPPARAM = bpparam())

annotateSpliceEventType(
  result,
  txdb,
  addSpliceType = TRUE,
  addUTRoverlap = TRUE,
  BPPARAM = bpparam()
)

flagBlacklistRegions(
  result,
  blacklist_regions = NULL,
  assemblyVersion = c("hg19", "hg38")
)
}
\arguments{
\item{fds}{A FraserDataSet}

\item{txdb}{A txdb object providing the reference annotation.}

\item{BPPARAM}{For controlling parallelization behavior. Defaults to 
\code{bpparam()}.}

\item{result}{A result table as generated by FRASER, including the column 
\code{annotatedJunction} as generated by the function
\code{annotateIntronReferenceOverlap}.}

\item{addSpliceType}{Logical, indicating if the type of the splice event 
should be added to the results table. Defaults to \code{TRUE}.}

\item{addUTRoverlap}{Logical, indicating if the overlap with UTR regions
should checked and added to the results table. Defaults to \code{TRUE}.}

\item{blacklist_regions}{A BED file that contains the blacklist regions. 
If \code{NULL} (default), the BED files that are packaged with FRASER 
are used.}

\item{assemblyVersion}{Indicates the genome assembly version of the intron 
coordinates. Only used if blacklist_regions is NULL. For other versions, 
please provide the BED file containing the blacklist regions directly.}
}
\value{
An annotated FraserDataSet or results table, respectively
}
\description{
These functions work on the result table and add additional 
    annotations to the reported introns: the type of splice event (e.g. 
    exon skipping, exon truncation, ...), expected occurence of frameshift, 
    overlap with UTR regions as well as a flag for introns that are 
    located in blacklist regions of the genome.
    
\code{\link{annotateIntronReferenceOverlap}} adds basic annotations to the 
    fds for each intron based on the overlap of the intron's location with 
    the reference annotation. Has to be run before the result table is 
    created so that the new column can be included in it (see examples).

\code{\link{annotateSpliceEventType}} annotates each intron in the results 
    table with the type of splice event and expected occurence of frameshift 
    (likely, unlikely, inconclusive). Can also calculate overlap with 
    annotated UTR regions. Splice types can be one of: 
    annotatedIntron_increasedUsage, annotatedIntron_reducedUsage, 
    exonTruncation, exonElongation, exonTruncation&Elongation, 
    singleExonSkipping, exonSkipping, splicingBeyondGene, 
    multigenicSplicing, downstreamOfNearestGene, upstreamOfNearestGene, 
    complex (everything else).     
    Splice sites (theta metric) annotations indicate how the splice site is 
    located with respect to the reference annotation. The annotated types 
    are: annotatedSpliceSite, exonicRegion, intronicRegion.

\code{\link{flagBlacklistRegions}} flags introns in the results table on 
    whether or not they are located in a blacklist region of the genome. By
    default, the blacklist regions as reported in \cite{...} are used.
}
\section{Functions}{
\itemize{
\item \code{annotateIntronReferenceOverlap}: This method calculates basic annotations 
based on overlap with the reference annotation (start, end, none, both) 
for the full fds. The overlap type is added as a new column 
\code{annotatedJunction} in \code{mcols(fds)}.

\item \code{annotateSpliceEventType}: This method annotates the splice event 
type to junctions in the given results table.

\item \code{flagBlacklistRegions}: This method flags blacklist regions in 
the given results table.
}}

\examples{
  # get data, fit and compute p-values and z-scores
  fds <- createTestFraserDataSet()
  
  # load reference annotation
  requireNamespace("TxDb.Hsapiens.UCSC.hg19.knownGene")
  txdb <- 
    TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene 
    
  # add basic annotations for overlap with the reference annotation
  # run this function before creating the results table
  fds <- annotateIntronReferenceOverlap(fds, txdb) 

  # extract results: for this small example dataset, only a z score cutoff 
  # of 1 is used to get at least one result. 
  # Make sure to include the additional column in the results table
  res <- results(fds, padjCutoff=NA, zScoreCutoff=1, deltaPsiCutoff=NA, 
                  additionalColumns = 'annotatedJunction')
  res_dt <- as.data.table(res)
  
  # annotate the type of splice event and UTR overlap
  res_dt <- annotateSpliceEventType(result=res_dt, txdb=txdb)
  
  # annotate overlap with blacklist regions
  res_dt <- flagBlacklistRegions(result=res_dt, assemblyVersion="hg19")
  
  # show results table containing additional annotations
  res_dt
  
}
