% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{strandSpecific}
\alias{strandSpecific}
\alias{strandSpecific<-}
\alias{strandSpecific,FraseRDataSet-method}
\alias{strandSpecific<-,FraseRDataSet-method}
\title{Get/Set if the analysis is strand specific or not}
\usage{
strandSpecific(object)

strandSpecific(object) <- value

\S4method{strandSpecific}{FraseRDataSet}(object)

\S4method{strandSpecific}{FraseRDataSet}(object) <- value
}
\arguments{
\item{object}{A FraseRDataSet object.}

\item{value}{The new value for the strand specificity of the analysis.}
}
\value{
A logical value if the analysis is strand specific
}
\description{
Get/Set if the analysis is strand specific or not
}
\examples{
settings <- createTestFraseRSettings()
strandSpecific(settings)
strandSpecific(settings) <- TRUE
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
