% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{[,FraseRDataSet,ANY,ANY,ANY-method}
\alias{[,FraseRDataSet,ANY,ANY,ANY-method}
\title{Subsetting by indices for junctions}
\usage{
\S4method{[}{FraseRDataSet,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{A \code{FraseRDataSet} object}

\item{i}{A integer vector to subset the rows/ranges}

\item{j}{A integer vector to subset the columns/samples}
}
\value{
A subsetted \code{FraseRDataSet} object
}
\description{
Providing subsetting by indices through the single-bracket operator
}
