% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseRSettings-class.R
\docType{methods}
\name{[,FraseRSettings,numeric,ANY,ANY-method}
\alias{[,FraseRSettings,numeric,ANY,ANY-method}
\title{Subsetting by indices}
\usage{
\S4method{[}{FraseRSettings,numeric,ANY,ANY}(x, i)
}
\arguments{
\item{x}{A \code{FraseRSettings} object}

\item{i}{A integer vector}
}
\value{
A subsetted \code{FraseRSettings} object
}
\description{
Providing subsetting by indices through the single-bracket operator
}
