% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{subsetFraseR}
\alias{subsetFraseR}
\title{Subsetting by indices for junctions}
\usage{
subsetFraseR(x, i, j)
}
\arguments{
\item{x}{A \code{FraseRDataSet} object}

\item{i}{A integer vector to subset the rows/ranges}

\item{j}{A integer vector to subset the columns/samples}
}
\value{
A subsetted \code{FraseRDataSet} object
}
\description{
Providing subsetting by indices through the single-bracket operator
}
\examples{
    fds <- countRNAData(createTestFraseRSettings())
    fds[1:10,1:10]
    fds[,samples(fds) \%in\% c("sample1", "sample2")]

}
