% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{aberrant}
\alias{aberrant}
\title{Aberrant splicing events}
\usage{
aberrant(fds, type = currentType(fds), padjCutoff = 0.05,
  deltaPsiCutoff = 0.3, zScoreCutoff = NA, by = c(NA, "sample",
  "feature"), aggregate = FALSE, ...)
}
\arguments{
\item{fds}{FraseRDataSet object}

\item{type}{Splicing type}

\item{padjCutoff}{Adjusted p value cutoff to be used or NA if not requested}

\item{deltaPsiCutoff}{Delta PSI cutoff to be used or NA if not requested}

\item{zScoreCutoff}{Z score cutoff to be used or NA if not requested}

\item{by}{By default NA which means no grouping. But if \code{sample}
or \code{feature} is specified the sum by sample or feature is
returned}

\item{aggregate}{If TRUE the returned object is based on the grouped 
features}

\item{...}{Further arguments can be passed to the method. If "zscores", 
"padjVals" or "dPsi" is given, the values of those arguments are used to 
define the aberrant events.}
}
\value{
Either a of logical values of size introns/genes x samples if "by" 
is NA or a vector with the number of aberrant events per sample or feature 
depending on the vaule of "by"
}
\description{
Use cutoffs to define splicing outlier event.
}
\examples{
# TODO
TODO <- 1

}
