% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationOfRanges.R
\name{annotateRanges}
\alias{annotateRanges}
\alias{annotateRangesWithTxDb}
\title{Annotates the given FraserDataSet with the HGNC symbol with biomaRt}
\usage{
annotateRanges(
  fds,
  feature = "hgnc_symbol",
  featureName = feature,
  biotype = list("protein_coding"),
  ensembl = NULL,
  GRCh = 37
)

annotateRangesWithTxDb(
  fds,
  feature = "SYMBOL",
  featureName = "hgnc_symbol",
  keytype = "ENTREZID",
  txdb = NULL,
  orgDb = NULL
)
}
\arguments{
\item{fds}{FraserDataSet}

\item{feature}{Defines which feature (default is HGNC symbol) should be 
annotated.}

\item{featureName}{Name of the feature in the FraserDataSet mcols.}

\item{biotype}{The biotype.}

\item{ensembl}{The ensembl that should be used. If NULL, the default one is 
used (hsapiens_gene_ensembl, GRCh37).}

\item{GRCh}{GRCh version to connect to. If this is NULL, then the current 
GRCh38 is used. Otherwise, this can only be 37 (default) at the moment 
(see \code{useEnsembl}).}

\item{keytype}{The type of gene IDs in the \code{TxDb} object (see 
AnnotationDbi::keytypes(orgDb) for a list of available ID types).}

\item{txdb}{A \code{TxDb} object. If this is NULL, then the default 
one is used, currently this is \code{TxDb.Hsapiens.UCSC.hg19.knownGene}.}

\item{orgDb}{An \code{orgDb} object If this is NULL, then the 
default one is used, currently this is \code{org.Hs.eg.db}.}
}
\value{
FraserDataSet
}
\description{
Annotates the given FraserDataSet with the HGNC symbol with biomaRt
}
\examples{

fds <- createTestFraserDataSet()

\dontrun{
### Two ways to annotage ranges with gene names: 
# either using biomart:
fds <- annotateRanges(fds, GRCh=NULL)
rowRanges(fds, type="psi5")[,"hgnc_symbol"]
 
# or with a TxDb object
require(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
require(org.Hs.eg.db)
orgDb <- org.Hs.eg.db
fds <- annotateRangesWithTxDb(fds, txdb=txdb, orgDb=orgDb)

rowRanges(fds, type="psi5")[,"hgnc_symbol"]
}

}
