% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationOfRanges.R
\name{annotateRanges}
\alias{annotateRanges}
\title{Annotates the given FraseRDataSet with the HGNC symbol
with biomaRt}
\usage{
annotateRanges(dataset, feature = "hgnc_symbol",
  biotype = list("protein_coding"), ensembl = useEnsembl(biomart =
  "ensembl", dataset = "hsapiens_gene_ensembl", GRCh = 37))
}
\description{
Annotates the given FraseRDataSet with the HGNC symbol
with biomaRt
}
