% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{assayNames,FraseRDataSet-method}
\alias{assayNames,FraseRDataSet-method}
\title{Returns the assayNames of FraseR}
\usage{
\S4method{assayNames}{FraseRDataSet}(x)
}
\arguments{
\item{x}{FraseRDataSet}
}
\value{
Character vector
}
\description{
Returns the assayNames of FraseR
}
