% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{assays,FraserDataSet-method}
\alias{assays,FraserDataSet-method}
\alias{assays<-,FraserDataSet,SimpleList-method}
\alias{assays<-,FraserDataSet,list-method}
\alias{assays<-,FraserDataSet,DelayedMatrix-method}
\title{Returns the assay for the given name/index of the FraserDataSet}
\usage{
\S4method{assays}{FraserDataSet}(x, withDimnames = TRUE, ...)

\S4method{assays}{FraserDataSet,SimpleList}(x, withDimnames = TRUE, HDF5 = TRUE, type = NULL, ...) <- value

\S4method{assays}{FraserDataSet,list}(x, withDimnames = TRUE, HDF5 = TRUE, type = NULL, ...) <- value

\S4method{assays}{FraserDataSet,DelayedMatrix}(x, withDimnames = TRUE, HDF5 = TRUE, type = NULL, ...) <- value
}
\arguments{
\item{x}{FraserDataSet}

\item{withDimnames}{Passed on to SummarizedExperiment::assays()}

\item{...}{Parameters passed on to SummarizedExperiment::assays()}

\item{HDF5}{Logical value indicating whether the assay should be stored as 
a HDF5 file.}

\item{type}{The psi type.}

\item{value}{The new value to which the assay should be set.}
}
\value{
(Delayed) matrix.
}
\description{
Returns the assay for the given name/index of the FraserDataSet
}
