% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{assays,FraseRDataSet-method}
\alias{assays,FraseRDataSet-method}
\alias{assays<-,FraseRDataSet,SimpleList-method}
\alias{assays<-,FraseRDataSet,list-method}
\alias{assays<-,FraseRDataSet,DelayedMatrix-method}
\title{Returns the assay for the given name/index of the FraseRDataSet}
\usage{
\S4method{assays}{FraseRDataSet}(x, withDimnames = TRUE, ...)

\S4method{assays}{FraseRDataSet,SimpleList}(x, withDimnames = TRUE, HDF5 = TRUE, type = NULL, ...) <- value

\S4method{assays}{FraseRDataSet,list}(x, withDimnames = TRUE, HDF5 = TRUE, type = NULL, ...) <- value

\S4method{assays}{FraseRDataSet,DelayedMatrix}(x, withDimnames = TRUE, HDF5 = TRUE, type = NULL, ...) <- value
}
\arguments{
\item{x}{FraseRDataSet}

\item{withDimnames}{Passed on to SummarizedExperiment::assays()}

\item{...}{Parameters passed on to SummarizedExperiment::assays()}

\item{HDF5}{Logical value indicating whether the assay should be stored as 
a HDF5 file.}

\item{type}{The psi type.}

\item{value}{The new value to which the assay should be set.}
}
\value{
(Delayed) matrix.
}
\description{
Returns the assay for the given name/index of the FraseRDataSet
}
