% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{assays,FraseRDataSet-method}
\alias{assays,FraseRDataSet-method}
\title{Returns the assay for the given name/index of the FraseRDataSet}
\usage{
\S4method{assays}{FraseRDataSet}(x, ..., withDimnames = TRUE)
}
\arguments{
\item{x}{FraseRDataSet}

\item{...}{Parameters passed on to SummarizedExperiment::assays()}

\item{withDimnames}{Passed on to SummarizedExperiment::assays()}
}
\value{
(Delayed) matrix.
}
\description{
Returns the assay for the given name/index of the FraseRDataSet
}
