% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{assays,FraseRDataSet-method}
\alias{assays,FraseRDataSet-method}
\title{Returns the assay corrensonding to the given name/index of the FraseRDataSet}
\usage{
\S4method{assays}{FraseRDataSet}(x, ..., withDimnames = TRUE)
}
\value{
(Delayed) matrix.
}
\description{
Returns the assay corrensonding to the given name/index of the FraseRDataSet
}
