% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePSIValue.R
\name{calculatePSIValues}
\alias{calculatePSIValues}
\title{This function calculates the PSI values for each junction and splice site
based on the FraseRDataSet object}
\usage{
calculatePSIValues(dataset)
}
\description{
This function calculates the PSI values for each junction and splice site
based on the FraseRDataSet object
}
\examples{
  fds <- counRNAData(createTestFraseRSettings())
  fds <- calculatePSIValues(fds)
}
