% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStats.R
\name{calculateZScores}
\alias{calculateZScores}
\title{Calculate the zscore for each PSI value.}
\usage{
calculateZScores(fds, type = psiTypes)
}
\description{
Calculate the zscore for each PSI value.
}
\examples{
  fds <- countRNAData(createTestFraseRSettings())
  fds <- calculatePSIValues(fds)
  fds <- calculateZScores(fds)
}
