% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countRNAseqData.R
\name{countRNA}
\alias{countRNA}
\alias{countRNAData}
\alias{getSplitReadCountsForAllSamples}
\alias{getNonSplitReadCountsForAllSamples}
\alias{addCountsToFraseRDataSet}
\alias{countSplitReads}
\alias{mergeCounts}
\alias{countNonSplicedReads}
\title{Count RNA-seq data}
\usage{
countRNAData(fds, NcpuPerSample = 1, junctionMap = NULL,
  minAnchor = 5, recount = FALSE, BPPARAM = bpparam(),
  genome = NULL, countDir = file.path(workingDir(fds), "savedObjects",
  nameNoSpace(name(fds))), ...)

getSplitReadCountsForAllSamples(fds, NcpuPerSample = 1,
  junctionMap = NULL, recount = FALSE, BPPARAM = bpparam(),
  genome = NULL, countFiles = NULL)

getNonSplitReadCountsForAllSamples(fds, splitCounts, NcpuPerSample = 1,
  minAnchor = 5, recount = FALSE, BPPARAM = bpparam(),
  longRead = TRUE)

addCountsToFraseRDataSet(fds, splitCounts, nonSplitCounts)

countSplitReads(sampleID, fds, NcpuPerSample = 1, genome = NULL,
  recount = FALSE)

mergeCounts(countList, junctionMap = NULL, assumeEqual = FALSE,
  BPPARAM = SerialParam())

countNonSplicedReads(sampleID, splitCounts, fds, NcpuPerSample = 1,
  minAnchor = 5, recount = FALSE, longRead = TRUE)
}
\arguments{
\item{fds}{A FraseRDataSet object}

\item{NcpuPerSample}{A BiocParallel param object or a positive integer
to configure the parallel backend
of the internal loop per sample}

\item{junctionMap}{A object or file containing a map of
all junctions of interest across all samples}

\item{minAnchor}{Minimum overlap around the Donor/Acceptor for
non spliced reads. Default to 5}

\item{recount}{if TRUE the cache is ignored and the bam file is recounted.}

\item{BPPARAM}{the BiocParallel parameters for the parallelization}

\item{genome}{the genome}

\item{countDir}{The directory in which the tsv containing the position and 
counts of the junctions should be placed.}

\item{...}{Further parameters passed on to Rsubread::featureCounts.}

\item{countFiles}{If specified, the split read counts for all samples are 
read from the specified files. Should be a vector of paths 
to files containing the split read counts for the 
individual samples. Reading from files is only supported 
for tsv(.gz) or RDS files containing GRranges objects. The 
order of the individual sample files should correspond to 
the order of the samples in the fds.}

\item{splitCounts}{The merged GRanges object containing the position and 
counts of all the introns in the dataset for all samples.}

\item{longRead}{If TRUE, then the isLongRead option of 
Rsubread::featureCounts is used when counting the non spliced reads 
overlapping splice sites.}

\item{nonSplitCounts}{The merged GRanges object containing the position and 
non split read counts of all splice sites present in the 
dataset for all samples.}

\item{sampleID}{The ID of the sample to be counted.}

\item{countList}{A list of GRanges objects containing the counts that should
be merged into one object.}

\item{assumeEqual}{Logical indicating whether all objects in 
\code{countList} can be assumed to contain counts for the same ranges. 
If FALSE, merging of the ranges is performed.}
}
\value{
\code{\link{countRNAData}} returns a FraseRDataSet.

\code{\link{getSplitReadCountsForAllSamples}} returns a GRanges 
object.

\code{\link{getNonSplitReadCountsForAllSamples}} returns a 
         GRanges object.

\code{\link{addCountsToFraseRDataSet}} returns a FraseRDataSet.

\code{\link{countSplitReads}} returns a GRanges object.

\code{\link{mergeCounts}} returns a GRanges object.

\code{\link{countNonSplicedReads}} returns a GRanges object.
}
\description{
The FraseR package provides multiple functions to extract and 
              count both split and non-spliced reads from bam files. 
              See Detail and Functions for more information.
}
\details{
The functions described in this file extract and count both the 
split and the non-spliced reads from bam files. 

\code{\link{countRNAData}} is the main function that takes care of all 
counting steps and returns a FraseRDataSet containing the counts for all 
samples in the fds. 

\code{\link{getSplitReadCountsForAllSamples}} counts split reads for all 
samples and \code{\link{getNonSplitReadCountsForAllSamples}} counts non 
split reads overlapping splice sites for all samples. 
\code{\link{addCountsToFraseRDataSet}} adds these counts to an existing fds.

\code{\link{countSplitReads}} calculates the split read counts for a single
sample. \code{\link{countNonSplicedReads}} counts the non split reads 
overlapping with splice sites for a single sample.

\code{\link{mergeCounts}} merges the counts from different samples into a 
single count object, where the counts for junctions that are not present in 
a sample are set to zero.
}
\section{Functions}{
\itemize{
\item \code{countRNAData}: This method extracts and counts the split reads and
non spliced reads from RNA bam files.

\item \code{getSplitReadCountsForAllSamples}: This method creates a GRanges object containing the 
split read counts from all specified samples.

\item \code{getNonSplitReadCountsForAllSamples}: This method creates a GRanges object containing the 
non split read counts at the exon-intron boundaries inferred 
from the GRanges object containing the positions of all the 
introns in this dataset.

\item \code{addCountsToFraseRDataSet}: This method adds the split read and non split read
counts to a existing FraseRDataSet containing the settings.

\item \code{countSplitReads}: This method counts all split reads in a bam file for a 
single sample.

\item \code{mergeCounts}: This method merges counts for multiple samples into one 
SummarizedExperiment object.

\item \code{countNonSplicedReads}: This method counts non spliced reads based on the 
given target (acceptor/donor) regions for a single sample.
}}

\examples{
  fds <- countRNAData(createTestFraseRSettings())
  fds <- countRNAData(createTestFraseRSettings(), 5)

}
