% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countRNAseqData.R
\name{countRNAData}
\alias{countRNAData}
\title{Count RNA-seq data}
\usage{
countRNAData(fds, NcpuPerSample = 1, junctionMap = NULL,
  minAnchor = 5, recount = FALSE, BPPARAM = bpparam(),
  genome = NULL)
}
\arguments{
\item{fds}{A FraseRDataSet object with all the information
how and what to count}

\item{NcpuPerSample}{A BiocParallel param object or a positive integer
to configure the parallel backend
of the internal loop per sample}

\item{junctionMap}{A object or file containing a map of
all junctions of interest across all samples}

\item{minAnchor}{Minimum overlap around the Donor/Acceptor for
non spliced reads. Default to 5}

\item{recount}{if TRUE the cache is ignored and the bam file is recounted.}
}
\value{
FraseRDataSet
}
\description{
This method extracts and counts the split reads and
            non spliced reads from a RNA bam file
}
\details{
TODO
}
\examples{
  countRNAData(createTestFraseRSettings())
  countRNAData(createTestFraseRSettings(), 5)
}
