% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{counts,FraseRDataSet-method}
\alias{counts,FraseRDataSet-method}
\title{getter for count data}
\usage{
\S4method{counts}{FraseRDataSet}(object, type = NULL, side = c("ofInterest",
  "otherSide"))
}
\description{
getter for count data
}
