% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNSetterFuns.R, R/AllGenerics.R
\docType{methods}
\name{K}
\alias{K}
\alias{N}
\alias{counts,FraseRDataSet-method}
\alias{counts<-,FraseRDataSet,ANY-method}
\title{Getter/setter for count data}
\usage{
K(fds, type = currentType(fds))

N(fds, type = currentType(fds))

\S4method{counts}{FraseRDataSet}(object, type = NULL,
  side = c("ofInterest", "otherSide"))

\S4method{counts}{FraseRDataSet,ANY}(object, type = NULL,
  side = c("ofInterest", "otherSide"), ...) <- value
}
\arguments{
\item{fds, object}{FraseRDataSet}

\item{type}{The psi type.}

\item{side}{"ofInterest" for junction counts, "other" for sum of counts of 
all other junctions at the same donor site (psi5) or acceptor site (psi3), 
respectively.}

\item{value}{An integer matrix containing the counts.}
}
\value{
FraseRDataSet
}
\description{
Getter/setter for count data

setter for count data
}
\examples{
 fds <- makeExampleFraseRDataSet()
 counts(fds, type="psi5", side="ofInterest")
 counts(fds, type="psi5", side="other")
 head(K(fds))
 head(N(fds))
 
}
