% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{createTestFraseRSettings}
\alias{createTestFraseRSettings}
\alias{createTestFraseRDataSet}
\title{Create a test dataset}
\usage{
createTestFraseRSettings(workingDir = tempdir())

createTestFraseRDataSet(workingDir = tempdir(), rerun = FALSE)
}
\arguments{
\item{workingDir}{directory where to store HDF5 and RDS files. Defaults to
the current tempory R session folder.}

\item{rerun}{Defaults to \code{FALSE}. If set to \code{TRUE} it reruns the
full fit of the model.}
}
\value{
a FraseRDataSet object which contains a test case
}
\description{
Create a test case dataset based on the test sample annotation to be 
used in the vignette and to explore the functionallity of the 
FRASER package. Dependent on the request only the sample annotation 
or a full fitted model is returned.
}
\examples{
fds <- createTestFraseRSettings()
fds

fds <- createTestFraseRSettings("FRASER-data")
fds

fds <- createTestFraseRDataSet()
fds

}
