% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{createTestFraseRDataSet}
\alias{createTestFraseRDataSet}
\title{Create a test case dataset based on the test sample annotation
filled with counts to be used in the vignette and
to explore the functionallity of the FraseR package.}
\usage{
createTestFraseRDataSet(BPPARAM = NULL)
}
\value{
a FraseRDataSet object which contains a test case
}
\description{
Create a test case dataset based on the test sample annotation
filled with counts to be used in the vignette and
to explore the functionallity of the FraseR package.
}
\examples{
    createTestFraseRDataSet()
    
}

