% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterExpression.R
\name{filterExpression}
\alias{filterExpression}
\title{Filtering FraseRDataSets}
\usage{
filterExpression(fds, minExpressionInOneSample = 20, quantile = 0.05,
  quantileMinExpression = 1, minDeltaPsi = 0, filter = TRUE,
  BPPARAM = bpparam(), delayed = FALSE)
}
\description{
Filtering FraseRDataSets
}
