% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCorrectionMethods.R
\name{fit}
\alias{fit}
\title{Fit a beta-binomial distribution and correction for confounders}
\usage{
fit(
  fds,
  correction = c("PCA", "PCA-regression", "PCA-BB-Decoder", "FraseR", "FraseR-weighted",
    "PCA-BB-full", "PCA-reg-full", "FraseR-5DecoderBatches", "FraseR-1DecoderBatches",
    "fullFraseR", "PCA-BB-Decoder-no-weights", "BB"),
  q,
  type = "psi3",
  rhoRange = c(1e-08, 1 - 1e-08),
  weighted = FALSE,
  noiseAlpha = 1,
  convergence = 1e-05,
  iterations = 15,
  initialize = TRUE,
  control = list(),
  BPPARAM = bpparam(),
  nSubset = 15000,
  verbose = FALSE,
  minDeltaPsi = 0.1
)
}
\arguments{
\item{fds}{A FraseRDataSet object}

\item{correction}{The method that should be used to correct for confounders.}

\item{q}{The dimension of the latent space. Should be fitted using 
\code{\link{optimHyperParams}} if unknown.}

\item{type}{The type of PSI (psi5, psi3 or psiSite for theta/splicing 
efficiency)}

\item{rhoRange}{Defines the range of values that rho parameter from the 
beta-binomial distribution is allowed to take. For very small values of rho, 
the loss can be instable, so it is not recommended to allow rho < 1e-8.}

\item{weighted}{If TRUE, the weighted implementation of the autoencoder is 
used}

\item{noiseAlpha}{Controls the amount of noise that is added for the 
denoising autoencoder.}

\item{convergence}{The fit is considered to have converged if the difference 
between the previous and the current loss is smaller than this threshold.}

\item{iterations}{The maximal number of iterations. When the autoencoder has 
not yet converged after these number of iterations, the fit stops anyway.}

\item{initialize}{If FALSE and a fit has been previoulsy run, the values 
from the previous fit will be used as initial values. If TRUE, 
(re-)initialization will be done.}

\item{control}{List of control parameters passed on to optim().}

\item{BPPARAM}{the BiocParallel parameters for the parallelization}

\item{nSubset}{The size of the subset to be used in fitting if subsetting is
used.}

\item{verbose}{Controls the level of information printed during the fit.}

\item{minDeltaPsi}{Minimal delta psi of an intron to be be considered a 
variable intron.}
}
\value{
FraseRDataSet
}
\description{
This method correct for confounders in the data and fits a beta-binomial 
distribution to the introns.
}
\details{
Available methods to correct for the confounders are currently: a denoising 
autoencoder with a BB loss ("FraseR"), PCA ("PCA"), a hybrid approach where 
PCA is used to fit the latent space and then the decoder of the autoencoder 
is fit using the BB loss ("PCA-BB-Decoder"). Although not recommended, it is 
also possbile to directly fit the BB distrbution to the raw counts ("BB").
}
\examples{
  # generate toy data
  fds <- makeExampleFraseRDataSet()
  fds <- calculatePSIValues(fds)
  fds <- filterExpressionAndVariability(fds) 
  
  # fit
  fds <- fit(fds, correction="PCA", q=3, type="psi5")
  
}
