% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRASER-package.R
\name{FRASER}
\alias{FRASER}
\title{FRASER: A package providing a workflow to detect aberrant splicing events
    in RNA sequencing data in a rare disease cohort}
\description{
FRASER: A package providing a workflow to detect aberrant splicing events
    in RNA sequencing data in a rare disease cohort
}
