% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseR-package.R, R/Fraser-pipeline.R
\name{FraseR}
\alias{FraseR}
\title{FraseR: A package providing a workflow to detect aberrant splicing events
    in RNA sequencing data in a rare disease cohort}
\usage{
FraseR(fds, q, NcpuPerSample = 1, ...)
}
\arguments{
\item{NcpuPerSample}{A BiocParallel param object to configure the
parallel backend of the internal loop for counting}

\item{settings}{A FraseRDataSet object with all the information
how and what to count}
}
\value{
FraseRDataSet
}
\description{
This function performs a default analysis of RNA-seq data
}
\examples{
  fds <- createTestFraseRSettings()
  fds <- FraseR(fds)

  # save the final FraseR object
  saveFraseRDataSet(fds)

  # finally visualize the results
  plotSampleResults(fds, 'sample1')

}
