% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseR-package.R, R/Fraser-pipeline.R
\name{FraseR}
\alias{FraseR}
\title{FraseR: A package providing a workflow to detect aberrant splicing events
    in RNA sequencing data in a rare disease cohort}
\usage{
FraseR(fds, q, correction = "PCA", iterations = 15, BPPARAM = bpparam(), ...)
}
\arguments{
\item{fds}{A FraseRDataSet object}

\item{q}{The encoding dimensions to be used during the fitting proceadure.
If a named vector is provided it is used for the different
splicing types.}

\item{correction}{The method that should be used to correct for confounders.}

\item{iterations}{The maximal number of iterations. When the autoencoder has 
not yet converged after these number of iterations, the fit stops anyway.}

\item{BPPARAM}{A BiocParallel object to run the computation in parallel}

\item{...}{Additional parameters passed on to the internal fit function}
}
\value{
FraseRDataSet
}
\description{
The FraseR function runs the default FraseR pipeline combinig the fit,
the computation of Z scores and p values as well as the delta-PSI values.
All computed values are returned as an FraseRDataSet object. To have
more control over each analysis step, one can call each function separately.
}
\details{
* controlForConfounders to control for confounding effects
* fitParams to fit the additional beta binomial model parameters
        (only needed if the autoencoder is not used)
* computePvalues to calculate the nominal and adjusted p values
* computeZscores to calculate the Z scores
* computeDeltaPsi to calculate the delta PSI values
}
\examples{
  # preprocessing
  fds <- createTestFraseRSettings()
  fds <- countRNAData(fds)
  fds <- calculatePSIValues(fds)
  fds <- filterExpression(fds)

  # Run analysis pipeline
  fds <- FraseR(fds, q=2)
  fds

  # save the final FraseR object
  saveFraseRDataSet(fds)

  # finally visualize the results
  plotVolcano(fds, 'sample1', 'psi5')

}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
