% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraseR-package.R
\name{FraseR}
\alias{FraseR}
\title{FraseR: A package providing a workflow to detect aberrant splicing events
    in RNA sequencing data in a rare disease cohort}
\description{
FraseR: A package providing a workflow to detect aberrant splicing events
    in RNA sequencing data in a rare disease cohort
}
