% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraserDataSet-class.R
\docType{class}
\name{FraserDataSet-class}
\alias{FraserDataSet-class}
\title{FraserDataSet}
\description{
This class is designed to store the whole FRASER data set
needed for an analysis of a disease cohort
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
