% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNSetterFuns.R
\name{getter_setter_functions}
\alias{getter_setter_functions}
\alias{zScore,}
\alias{pVals,}
\alias{padjVals,}
\alias{rho,}
\alias{bestQ}
\alias{rho}
\alias{zScores}
\alias{pVals}
\alias{padjVals}
\alias{predictedMeans}
\alias{deltaPsiValue}
\alias{hyperParams}
\alias{dontWriteHDF5}
\alias{dontWriteHDF5<-}
\title{Getter/Setter functions}
\usage{
rho(fds, type = currentType(fds))

zScores(fds, type = currentType(fds), byGroup = FALSE, ...)

pVals(fds, type = currentType(fds), dist = "BetaBinomial", ...)

padjVals(fds, type = currentType(fds), dist = c("BetaBinomial"), ...)

predictedMeans(fds, type = currentType(fds))

deltaPsiValue(fds, type = currentType(fds))

hyperParams(fds, type = currentType(fds), all = FALSE)

bestQ(fds, type = currentType(fds))

dontWriteHDF5(fds)

dontWriteHDF5(fds) <- value
}
\arguments{
\item{fds}{An FraseRDataSet object.}

\item{type}{The type of psi (psi5, psi3 or psiSite)}

\item{byGroup}{If TRUE, aggregation by donor/acceptor site will be done.}

\item{...}{Internally used parameteres.}

\item{dist}{Distribution for which the p-values should be extracted.}

\item{all}{Logical value indicating whether \code{hyperParams(fds)} should 
return the results of all evaluated parameter combinations or only 
for the optimal parameter combination.}

\item{value}{The new value to be assigned.}
}
\value{
A (delayed) matrix or vector dependent on the type of data retrieved.
}
\description{
This is a collection of small accessor/setter functions for easy access to
the values within the FraseR model.
}
\section{Functions}{
\itemize{
\item \code{rho}: Returns the fitted rho values for the 
beta-binomial distribution

\item \code{zScores}: This returns the calculated z-scores.

\item \code{pVals}: This returns the calculated p-values.

\item \code{padjVals}: This returns the adjusted p-values.

\item \code{predictedMeans}: This returns the fitted mu (i.e. psi) 
values.

\item \code{deltaPsiValue}: Returns the difference between the 
observed and the fitted psi values.

\item \code{hyperParams}: This returns the results of the 
hyperparameter optimization NULL if the hyperparameter 
opimization was not run yet.

\item \code{bestQ}: This returns the optimal size of the 
latent space according to the hyperparameter optimization or a simple 
estimate of about a tenth of the number of samples if the hyperparameter 
opimization was not run yet.

\item \code{dontWriteHDF5}: Gets the current value of whether the 
assays should be stored as hdf5 files.

\item \code{dontWriteHDF5<-}: Sets whether the assays should be stored 
as hdf5 files.
}}

\examples{
fds <- createTestFraseRDataSet()
dontWriteHDF5(fds)
dontWriteHDF5 <- TRUE
currentType(fds) <- "psi5"

bestQ(fds)
rho(fds)
# get statistics
pVals(fds)
padjVals(fds)
zScores(fds)

}
