% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{length,FraseRDataSet-method}
\alias{length,FraseRDataSet-method}
\title{retrive the length of the object (aka number of junctions)}
\usage{
\S4method{length}{FraseRDataSet}(x)
}
\value{
Length of the object.
}
\description{
retrive the length of the object (aka number of junctions)
}
