% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveHDF5Objects.R
\name{loadFraseRDataSet}
\alias{loadFraseRDataSet}
\title{Load a saved FraseR object into memory}
\usage{
loadFraseRDataSet(dir)
}
\arguments{
\item{dir}{a path to the working directory of FraseR}
}
\description{
Load a saved FraseR object into memory
}
