% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-makeExampleObj.R
\name{makeExampleFraseRDataSet}
\alias{makeExampleFraseRDataSet}
\title{Create example data sets for FraseR}
\usage{
makeExampleFraseRDataSet()
}
\value{
FraseRDataSet
}
\description{
Creates an example data set from a file or simulates a data set based
on  random split read counts following a beta-binomial distribution with
injected outliers.
}
\examples{
  fds <- makeExampleFraseRDataSet()
  fds

}
