% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-makeExampleObj.R
\name{makeExampleFraseRDataSet}
\alias{makeExampleFraseRDataSet}
\title{make an FraseR example data set
using an subset of the Kremer paper (200 junctions)}
\usage{
makeExampleFraseRDataSet()
}
\description{
make an FraseR example data set
using an subset of the Kremer paper (200 junctions)
}
