% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-makeExampleObj.R
\name{makeExampleFraseRDataSet}
\alias{makeExampleFraseRDataSet}
\alias{makeFittedExampleFraseRDataSet}
\title{Create example data sets for FraseR}
\usage{
makeExampleFraseRDataSet(workingDir = tempdir())

makeFittedExampleFraseRDataSet(workingDir = tempdir(), rerun = FALSE)
}
\value{
FraseRDataSet
}
\description{
Creates an example data set from files with example counts. 
\code{makeFittedExampleFraseRDataSet} additionally runs the FRASER 
pipeline.
}
\examples{
  fds <- makeExampleFraseRDataSet()
  fds
  
}
