% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{mapSeqlevels}
\alias{mapSeqlevels}
\title{Mapping of chromosome names}
\usage{
mapSeqlevels(fds, style = "UCSC", ...)
}
\description{
Mapping of chromosome names
}
\examples{

fds <- countRNAData(createTestFraseRSettings())

seqlevels(mapSeqlevels(fds, style="UCSC"))
seqlevels(mapSeqlevels(fds, style="Ensembl"))
seqlevels(mapSeqlevels(fds, style="dbSNP"))

}
