% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{method,FraseRDataSet-method}
\alias{method,FraseRDataSet-method}
\alias{method<-,FraseRDataSet-method}
\title{Get/Set the statistical method to use for P-value calculation
 from the FraseRDataSet object}
\usage{
\S4method{method}{FraseRDataSet}(object)

\S4method{method}{FraseRDataSet}(object) <- value
}
\arguments{
\item{object}{A FraseRDataSet object.}
}
\value{
A character string representing the statistical method
}
\description{
Get/Set the statistical method to use for P-value calculation
 from the FraseRDataSet object
}
\examples{
settings <- createTestFraseRSettings()
method(settings)
method(settings) <- "betaBin"
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
