% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{nonSplicedReads,FraseRDataSet-method}
\alias{nonSplicedReads,FraseRDataSet-method}
\alias{nonSplicedReads<-,FraseRDataSet-method}
\title{Getter/setter for the non spliced reads object within the FraseRDataSet 
object}
\usage{
\S4method{nonSplicedReads}{FraseRDataSet}(object)

\S4method{nonSplicedReads}{FraseRDataSet}(object) <- value
}
\arguments{
\item{object}{A FraseRDataSet object.}
}
\value{
RangedSummarizedExperiment (getter) or FraseRDataSet (setter)
}
\description{
Getter/setter for the non spliced reads object within the FraseRDataSet 
object
}
