% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{nonSplicedReads}
\alias{nonSplicedReads}
\alias{nonSplicedReads<-}
\alias{nonSplicedReads,FraseRDataSet-method}
\alias{nonSplicedReads<-,FraseRDataSet-method}
\title{Getter/setter for the non spliced reads object within the FraseRDataSet 
object}
\usage{
nonSplicedReads(object)

nonSplicedReads(object) <- value

\S4method{nonSplicedReads}{FraseRDataSet}(object)

\S4method{nonSplicedReads}{FraseRDataSet}(object) <- value
}
\arguments{
\item{object}{A FraseRDataSet object.}

\item{value}{A RangedSummarizedExperiment object containing the counts for 
the non spliced reads overlapping splice sites in the fds.}
}
\value{
RangedSummarizedExperiment (getter) or FraseRDataSet (setter)
}
\description{
Getter/setter for the non spliced reads object within the FraseRDataSet 
object
}
