% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_encoding_dimensions.R
\name{optimHyperParams}
\alias{optimHyperParams}
\title{remove non variable junctions}
\usage{
optimHyperParams(fds, type, q_param = seq(3, min(30, ncol(fds)), by = 3),
  noise_param = c(0, 0.5, 1, 2), minDeltaPsi = 0.1,
  BPPARAM = bpparam(), iterations = 5, setSubset = 0.1,
  injectFreq = 0.01)
}
\description{
remove non variable junctions
}
