% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_encoding_dimensions.R
\name{optimHyperParams}
\alias{optimHyperParams}
\title{Find optimal encoding dimension}
\usage{
optimHyperParams(fds, type, correction, q_param = seq(2, min(40,
  ncol(fds)), by = 3), noise_param = c(0, 0.5, 1, 2, 5),
  minDeltaPsi = 0.1, iterations = 5, setSubset = 15000,
  injectFreq = 0.01, nrDecoderBatches = 1, BPPARAM = bpparam(),
  internalThreads = 3)
}
\description{
Finds the optimal encoding dimension by injecting artificial splicing outlier
ratios while maximizing the precision-recall curve.
}
\examples{
# TODO
TODO <- 1

}
