% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{parallel,FraseRDataSet-method}
\alias{parallel,FraseRDataSet-method}
\alias{parallel<-,FraseRDataSet-method}
\title{Get/Set the parallel param object from the FraseRDataSet object}
\usage{
\S4method{parallel}{FraseRDataSet}(object)

\S4method{parallel}{FraseRDataSet}(object) <- value
}
\arguments{
\item{object}{A FraseRDataSet object.}
}
\value{
A parallel param object
}
\description{
Get/Set the parallel param object from the FraseRDataSet object
}
\examples{
settings <- createTestFraseRSettings()
parallel(settings)
parallel(settings) <- SerialParam()
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
