% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistributions.R
\name{plotCountCorHeatmap}
\alias{plotCountCorHeatmap}
\title{Plot count correlation}
\usage{
plotCountCorHeatmap(fds, type = c("psi5", "psi3", "psiSite"),
  logit = FALSE, topN = 50000, minMedian = 1, main = NULL,
  normalized = FALSE, show_rownames = FALSE, show_colnames = FALSE,
  minDeltaPsi = 0.1, annotation_col = NA, annotation_row = NA,
  border_color = NA, topJ = 5000, plotType = c("sampleCorrelation",
  "junctionSample"), nClust = 5, sampleClustering = NULL,
  plotMeanPsi = TRUE, plotCov = TRUE, ...)
}
\description{
Plot count correlation
}
