% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterExpression.R
\name{plotFilterExpression}
\alias{plotFilterExpression}
\title{Plot filter expression}
\usage{
plotFilterExpression(fds, bins = 200, legend.position = c(0.8, 0.8))
}
\description{
Histogram of the geometric mean per junction based on the filter status
}
