% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{plotFunctions}
\alias{plotFunctions}
\alias{plotAberrantPerSample}
\alias{plotVolcano}
\alias{plotQQ}
\alias{plotExpression}
\alias{plotCountCorHeatmap}
\alias{plotFilterExpression}
\alias{plotExpectedVsObservedPsi}
\alias{plotEncDimSearch}
\title{Visualization functions for FraseR}
\usage{
plotVolcano(fds, sampleID, type, deltaPsiCutoff = 0.3,
  padjCutoff = 0.05, basePlot = TRUE, aggregate = FALSE,
  main = paste0("Volcano plot: ", sampleID))

plotAberrantPerSample(fds, main, type = c("psi3", "psi5", "psiSite"),
  padjCutoff = 0.1, zScoreCutoff = NA, deltaPsiCutoff = 0.3,
  aggregated = TRUE, BPPARAM = bpparam(), ...)

plotExpression(fds, type = c("psi5", "psi3", "psiSite"), site = NULL,
  result = NULL, colGroup = NULL, basePlot = TRUE,
  main = paste0("Expression plot: ", site), ...)

plotExpectedVsObservedPsi(fds, type = c("psi5", "psi3", "psiSite"),
  idx = NULL, result = NULL, colGroup = NULL, main = NULL,
  basePlot = TRUE, padjCutoff = 0.05)

plotQQ(fds, type = NULL, idx = NULL, result = NULL,
  aggregate = FALSE, global = FALSE, main = NULL,
  conf.alpha = 0.05, samplePoints = 30000, basePlot = TRUE,
  Ncpus = min(4, getDTTable()), ...)

plotEncDimSearch(fds, type = c("psi3", "psi5", "psiSite"))

plotFilterExpression(fds, bins = 200, legend.position = c(0.8, 0.8))

plotCountCorHeatmap(fds, type = c("psi5", "psi3", "psiSite"),
  logit = FALSE, topN = 50000, topJ = 5000, minMedian = 1,
  main = NULL, normalized = FALSE, show_rownames = FALSE,
  show_colnames = FALSE, minDeltaPsi = 0.1, annotation_col = NA,
  annotation_row = NA, border_color = NA, nClust = 5,
  plotType = c("sampleCorrelation", "junctionSample"),
  sampleClustering = NULL, plotMeanPsi = TRUE, plotCov = TRUE, ...)
}
\arguments{
\item{fds, object}{An FraseRDataSet object.}

\item{sampleID}{A sample ID which should be plotted. Can also be a vector.
Integers are treated as indices.}

\item{type}{The psi type: either psi5, psi3 or psiSite (for SE).}

\item{padjCutoff, zScoreCutoff, deltaPsiCutoff}{Significance, Z-score or delta
psi cutoff to mark outliers}

\item{basePlot}{if \code{TRUE} (default), use the R base plot version, else
use the plotly framework.}

\item{aggregate}{If TRUE, the pvalues are aggregated by gene, otherwise
junction level pvalues are used (default).}

\item{main}{Title for the plot, if missing a default title will be used.}

\item{...}{Additional parameters passed to plot() or plot_ly() if not stated
otherwise in the details for each plot function}

\item{result}{The result table to be used by the method.}

\item{colGroup}{Group of samples that should be colored.}

\item{idx, site, feature}{A junction site ID or gene ID or one of both, which
should be plotted. Can also be a vector. Integers are treated
as indices.}

\item{global}{Flag to plot a global Q-Q plot, default FALSE}

\item{conf.alpha}{If set, a confidence interval is plotted, defaults to 0.05}

\item{samplePoints}{Sample points for Q-Q plot, defaults to max 30k points}

\item{bins}{Set the number of bins to be used in the histogram.}

\item{legend.position}{Set legend position (x and y coordinate), defaults to
the top right corner.}

\item{logit}{If TRUE, the default, psi values are plotted in logit space.}

\item{topN, topJ}{Top x most variable junctions that should be used in the
heatmap. TopN is used for sample-sample correlation heatmaps and
topJ for junction-sample correlation heatmaps.}

\item{minMedian, minDeltaPsi}{Minimal median value or minimal delta psi of a
junction to be considered for the correlation heatmap.}

\item{normalized}{If TRUE, the normalized psi values are used, the default,
otherwise the raw psi values}

\item{show_rownames, show_colnames}{Logical value indicating whether to show
row or column names on the heatmap axes.}

\item{annotation_col, annotation_row}{Row or column annotations that should be
plotted on the heatmap.}

\item{border_color}{Sets the border color of the heatmap}

\item{nClust}{Number of clusters to show in the row and
column dendrograms.}

\item{plotType}{Character string indicating the type of correlation heatmap
that should be plotted. Can be either 'sampleCorrelation' for a
sample-sample correlation heatmap or 'junctionSample' for a
junction-sample correlation heatmap.}

\item{sampleClustering}{A clustering of the samples that should be used as an
annotation of the heatmap.}

\item{plotMeanPsi, plotCov}{If \code{TRUE}, then the heatmap is annotated with
the mean psi values or the junction coverage.}

\item{col}{Set color for the barplot.}

\item{yadjust}{Option to adjust position of Median and 90 percentile labels.}

\item{ylab}{The y axis label}

\item{labCex}{The label cex parameter}

\item{labLine}{Option to move axis labels}

\item{breakTies}{If \code{TRUE},}

\item{highlightOutliers}{If \code{TRUE}, then outliers are highlighted.}

\item{maxOutlier}{Used to define the y axis limit.}

\item{cutYaxis}{If \code{TRUE}, cuts the y axis of the plot at some point.}

\item{ymax}{If set, ymax is the upper bound for the plot range on the y axis.}
}
\value{
If base R graphics are used nothing is returned else the plotly or
            the gplot object is returned.
}
\description{
The FraseR package provides mutliple functions to visualize
the data and the results of a full data set analysis.

Plots the p values over the delta psi values, known as volcano plot.
Visualizes per sample the outliers. By type and aggregate by
gene if requested.

Plot the number of aberrant events per samples

Plots the observed split reads of the junction of interest over all reads
coming from the given donor/acceptor.

Plots the expected psi value over the observed psi value of the given
junction.

Plots the quantile-quantile plot

Histogram of the geometric mean per junction based on the filter status

Count correlation heatmap function
}
\details{
This is the list of all plotting function provided by FraseR:
\itemize{
  \item plotAberrantPerSample()
  \item plotVolcano()
  \item plotExpression()
  \item plotQQ()
  \item plotExpectedVsObservedPsi()
  \item plotCountCorHeatmap()
  \item plotFilterExpression()
  \item plotEncDimSearch()
}

For a detailed description of each plot function please see the details.
Most of the functions share the same parameters.



\code{plotAberrantPerSample}: The number of aberrant events per sample are
plotted sorted by rank. The ... parameters are passed on to the
\code{\link{aberrant}} function.

\code{plotVolcano}: the volcano plot is sample-centric. It plots for a given
sample and psi type the negative log10 nominal P-values against the delta psi
values for all splice sites or aggregates by gene if requested.

\code{plotExpression}: This function plots for a given site the
read count at this site (i.e. K) against the total coverage (i.e. N) for the
given psi type (psi5, psi3 or SE (psiSite)) for all samples.

\code{plotQQ}: the quantile-quantile plot for a given gene or if
\code{global} is set to \code{TRUE} over the full data set. Here the
observed P-values are plotted against the expected ones in the negative
log10 space.

\code{plotExpectedVsObservedPsi}: A scatter plot of the observed psi
against the predicted psi for a given site.

\code{plotCountCorHeatmap}: The correlation heatmap of the count data either
of the full data set (i.e. sample-sample correlations) or of the top x most
variable junctions (i.e. junction-sample correlations). By default the values
are log transformed and row centered. The ... arguments are passed to the
\code{\link[pheatmap]{pheatmap}} function.

\code{plotFilterExpression}: The distribution of FPKM values. If the FraseRDataSet
object contains the \code{passedFilter} column, it will plot both FPKM
distributions for the expressed genes and for the filtered genes.

\code{plotEncDimSearch}: Visualization of the hyperparameter optimization.
It plots the encoding dimension against the achieved loss (area under the
precision-recall curve). From this plot the optimum should be choosen for
the \code{q} in fitting process.
}
\examples{
# TODO
TODO <- 1

}
