% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotSampleResults}
\alias{plotSampleResults}
\title{Plot the results of the FraseR analysis pipeline
All three types are plotted into one HTML file
based on plotly.}
\usage{
plotSampleResults(fds, sampleID = NULL, file = NULL, browseIt = FALSE)
}
\description{
Plot the results of the FraseR analysis pipeline
All three types are plotted into one HTML file
based on plotly.
}
\examples{
     fds <- FraseR()
     plotSampleResults(fds, "sample1")
     plotSampleResults(fds, "sample1", "result.html")
}
