% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotSampleResults}
\alias{plotSampleResults}
\title{Plot the results of the FraseR analysis pipeline
All three types are plotted into one HTML file
based on plotly.}
\usage{
plotSampleResults(dataset, sampleID = NULL, file = NULL)
}
\description{
Plot the results of the FraseR analysis pipeline
All three types are plotted into one HTML file
based on plotly.
}
\examples{
     plotSampleResults(dataset, "sample1")
     plotSampleResults(dataset, "sample1", "result.html")
}
