% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistributions.R
\name{plotValueVsCounts}
\alias{plotValueVsCounts}
\title{plot a given value against the counts}
\usage{
plotValueVsCounts(gr, fds, type, sample = NULL, delta = FALSE,
  main = paste0(type, " versus total raw counts"),
  xlab = "Number of all observed split reads", ylab = NULL,
  zeroVal = 0.64, rmZeroCts = FALSE, data = NULL, ...)
}
\description{
plot a given value against the counts
}
