% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot}
\usage{
plotVolcano(fds, sampleID, type, minDeltaPsi = 0.3, minPVal = 5.5,
  basePlot = TRUE, aggregate = FALSE)
}
\description{
Plots the p values over the delta psi values, known as volcano plot.
Visualizes per sample the outliers. By type and aggregate by
gene if requested.
}
