% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNSetterFuns.R
\name{pseudocount}
\alias{pseudocount}
\title{Set/get global pseudo count option}
\usage{
pseudocount(value = NULL)
}
\arguments{
\item{value}{If given, is sets the psuedocount. It requires a positive
number and rounds it to the next integer.}
}
\description{
Set and returns the pseudo count used within the FraseR fitting procedure.
}
\examples{
# set
pseudocount(4L)

# get
pseudocount()

}
