% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{psiTypes}
\alias{psiTypes}
\title{Available splice metrics}
\format{
An object of class \code{character} of length 4.
}
\usage{
psiTypes
}
\description{
Available splice metrics
}
\examples{
  # to show all available splice metrics:
  psiTypes

}
\keyword{datasets}
