% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{psiTypes}
\alias{psiTypes}
\title{Splice metrics that are run by default}
\format{
An object of class \code{character} of length 1.
}
\usage{
psiTypes
}
\description{
Splice metrics that are run by default
}
\examples{
  # to show splice metrics selected to be fitted:
  psiTypes

}
\keyword{datasets}
