% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{psiTypes}
\alias{psiTypes}
\title{Available psi types}
\format{An object of class \code{character} of length 3.}
\usage{
psiTypes
}
\description{
Available psi types
}
\examples{
  # to show available psi types:
  psiTypes

}
\keyword{datasets}
