% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{results}
\alias{results}
\title{obtain the results for the given analysis pipeline}
\usage{
results(fds, sampleIDs = samples(fds), pvalueCut = 1e-05, zscoreCut = 2,
  redo = FALSE)
}
\description{
obtain the results for the given analysis pipeline
}
