% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{samples,FraseRDataSet-method}
\alias{samples,FraseRDataSet-method}
\alias{samples<-,FraseRDataSet-method}
\title{Getter/Setter for the sampleIDs}
\usage{
\S4method{samples}{FraseRDataSet}(object)

\S4method{samples}{FraseRDataSet}(object) <- value
}
\arguments{
\item{object}{A FraseRDataSet object.}
}
\value{
A \code{vector} with all sample IDs
}
\description{
Getter/Setter for the sampleIDs
}
\examples{
fds <- createTestFraseRSettings()
samples(fds)
samples(fds) <- 1:length(fds)
}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
