% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{scanBamParam,FraseRDataSet-method}
\alias{scanBamParam,FraseRDataSet-method}
\alias{scanBamParam<-,FraseRDataSet-method}
\title{Get/Set the ScanBamParam object from the FraseRDataSet object}
\usage{
\S4method{scanBamParam}{FraseRDataSet}(object)

\S4method{scanBamParam}{FraseRDataSet}(object) <- value
}
\arguments{
\item{object}{A FraseRDataSet object.}
}
\value{
A ScanBamParam object
}
\description{
Get/Set the ScanBamParam object from the FraseRDataSet object
}
\examples{
settings <- createTestFraseRSettings()
scanBamParam(settings)
scanBamParam(settings) <- ScanBamParam(mapqFilter=30)

}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
