% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{subset.FraseR}
\alias{subset.FraseR}
\alias{[,FraseRDataSet,ANY,ANY,ANY-method}
\title{Subsetting by indices for junctions}
\usage{
\method{subset}{FraseR}(x, i, j, by = c("j", "ss"))

\S4method{[}{FraseRDataSet,ANY,ANY,ANY}(x, i, j, by = c("j", "ss"))
}
\arguments{
\item{x}{A \code{FraseRDataSet} object}

\item{i}{A integer vector to subset the rows/ranges}

\item{j}{A integer vector to subset the columns/samples}

\item{by}{a character (j or ss) definig if we subset by
junctions or splice sites}
}
\value{
A subsetted \code{FraseRDataSet} object
}
\description{
Providing subsetting by indices through the single-bracket operator
}
\examples{
    fds <- countRNAData(createTestFraseRSettings())
    fds[1:10,1:10]
    fds[,samples(fds) \%in\% c("sample1", "sample2")]
    fds[1:10,by="ss"]

}
