% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{workingDir,FraseRDataSet-method}
\alias{workingDir,FraseRDataSet-method}
\alias{workingDir<-,FraseRDataSet-method}
\title{Get/Set the working directory from the FraseRDataSet object}
\usage{
\S4method{workingDir}{FraseRDataSet}(object)

\S4method{workingDir}{FraseRDataSet}(object) <- value
}
\arguments{
\item{object}{A FraseRDataSet object.}

\item{value}{The new working directory.}
}
\value{
A path
}
\description{
Get/Set the working directory from the FraseRDataSet object
}
\examples{
settings <- createTestFraseRSettings()
workingDir(settings)
workingDir(settings) <- tempdir()

}
\author{
Christian Mertes \email{mertes@in.tum.de}
}
